+++ custom_components/southern_company/const.py
@@ -1,0 +1,3 @@
+"""Constants for the Southern Company integration."""
+
+DOMAIN = "southern_company"
